+++ shared-libs/ts-modules/start-core/lib/util/index.ts
@@ -1,6 +1,6 @@
 export {
   addressHostToUrl,
-  mdnsResolvable,
+  isAddressEnabled,
   filledAddress,
   filterNonLocal,
 } from './filledAddress'